requirements.txt
flask==2.3.3
opensearch-py==2.3.1
python-dotenv==1.0.0
faker==19.3.1
werkzeug==2.3.7
